+++ src/redux/actions/confirm-delete-access-type.js
@@ -1,0 +1,7 @@
+export const CONFIRM_DELETE_ACCESS_TYPE = 'CONFIRM_DELETE_ACCESS_TYPE';
+
+export function confirmDeleteAccessType() {
+  return {
+    type: CONFIRM_DELETE_ACCESS_TYPE,
+  };
+}